+++ app/Http/Livewire/Settings/CriticalDeviationMenu.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire\Settings;
 
 use Livewire\Component;
+use Illuminate\Support\Facades\Gate;
 use App\Models\CriticalDeviationMenu as CriticalDeviationMenuModel;
 use App\Models\CriticalDeviation as CriticalDeviationModel;
 use App\Models\Dropdown as DropdownModel;
@@ -31,6 +32,9 @@ class CriticalDeviationMenu extends Component
     public $limit = 10;
     public function mount($critical_deviation_id = null)
     {
+        if (!Gate::allows('allow-view', 'module-critical-deviation-management')) {
+            return redirect()->route('dashboard');
+        }
         $this->deviation = CriticalDeviationModel::find($critical_deviation_id);
     }
     public function render()
@@ -61,6 +65,15 @@ class CriticalDeviationMenu extends Component
     }
     public function onSave()
     {
+        $access = 'allow-create';
+        if($this->critical_deviation_menu_id){
+            $access = 'allow-edit';
+        }
+        if(!Gate::allows($access,'module-critical-deviation-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
+
         $this->validate([
             'label' => 'required',
             'remarks' => 'nullable|boolean',
@@ -93,6 +106,10 @@ class CriticalDeviationMenu extends Component
     }
     public function onDelete($id)
     {
+        if(!Gate::allows('allow-delete','module-critical-deviation-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
         $data = CriticalDeviationMenuModel::find($id);
         $data->delete();
     }
