+++ app/Http/Livewire/Settings/Dropdown.php
@@ -2,6 +2,7 @@
 namespace App\Http\Livewire\Settings;
 
 use Livewire\Component;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Dropdown as DropdownModel;
 use Livewire\WithPagination;
 use App\Helpers\CustomHelper;
@@ -17,6 +18,11 @@ class Dropdown extends Component
     public $modalTitle;
     public $modalButtonText;
     public $limit = 10;
+    public function mount(){
+        if (!Gate::allows('allow-view', 'module-dropdown-management')) {
+            return redirect()->route('dashboard');
+        }
+    }
     public function render()
     {
         $searchTerm = '%' . $this->searchTerm . '%';
@@ -27,6 +33,14 @@ class Dropdown extends Component
     }
     public function onSave()
     {
+        $access = 'allow-create';
+        if($this->dropdown_id){
+            $access = 'allow-edit';
+        }
+        if(!Gate::allows($access,'module-dropdown-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
         $this->validate(
             [
                 'name' => 'required',
@@ -53,6 +67,10 @@ class Dropdown extends Component
     }
     public function onDelete($dropdown_id)
     {
+        if(!Gate::allows('allow-delete','module-dropdown-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
         $dropdown = DropdownModel::find($dropdown_id);
         $dropdown->delete();
     }
