+++ app/Http/Livewire/Settings/Sanitary.php
@@ -1,5 +1,6 @@
 <?php
 namespace App\Http\Livewire\Settings;
+use Illuminate\Support\Facades\Gate;
 use App\Models\SanitaryModel;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -16,7 +17,12 @@ class Sanitary extends Component
     public $modalTitle;
     public $modalButtonText;
     public $limit = 10;
-    public function render()
+
+    public function mount(){
+        if (!Gate::allows('allow-view', 'module-sanitation-defect-management')) {
+            return redirect()->route('dashboard');
+        }
+    }    public function render()
     {
         $searchTerm = '%' . $this->searchTerm . '%';
         $sanitary_list = SanitaryModel::select('id', 'title', 'code')
@@ -36,6 +42,15 @@ class Sanitary extends Component
     }
     public function onSave()
     {
+        $access = 'allow-create';
+        if($this->sanitary_id){
+            $access = 'allow-edit';
+        }
+        if(!Gate::allows($access,'module-sanitation-defect-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
+
         $this->validate(
             [
                 'title' => 'required',
@@ -55,6 +70,10 @@ class Sanitary extends Component
     }
     public function onDelete($sanitary_id)
     {
+        if(!Gate::allows('allow-delete','module-sanitation-defect-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
         $sanitary = SanitaryModel::find($sanitary_id);
         $sanitary->delete();
     }
