+++ app/Http/Livewire/Settings/SubCategoryLabel.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Http\Livewire\Settings;
 use Livewire\Component;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Category as CategoryModel;
 use App\Models\SubCategory as SubCategoryModel;
 use App\Models\SubCategoryLabel as SubCategoryLabelModel;
@@ -29,6 +30,10 @@ class SubCategoryLabel extends Component
     public $limit = 10;
     public function mount($category_id = null, $sub_category_id = null)
     {
+
+        if (!Gate::allows('allow-view', 'module-category-management')) {
+            return redirect()->route('dashboard');
+        }
         $this->category_id = $category_id;
         $this->sub_category_id = $sub_category_id;
         $this->category_name = CategoryModel::where('id', $category_id)->value('name');
@@ -66,6 +71,14 @@ class SubCategoryLabel extends Component
     }
     public function onSave()
     {
+        $access = 'allow-create';
+        if($this->label_id){
+            $access = 'allow-edit';
+        }
+        if(!Gate::allows($access,'module-category-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
         SubCategoryLabelModel::updateOrCreate(
             ['id' => $this->label_id],
             [
@@ -82,6 +95,10 @@ class SubCategoryLabel extends Component
     }
     public function onDelete($id)
     {
+        if(!Gate::allows('allow-delete','module-category-management')){
+            $this->onAlert(false, 'Action Cancelled', 'Unable to perform action due to user is unauthorized!', 'warning');
+            return;
+        }
         $sub_category = SubCategoryLabelModel::find($id);
         $sub_category->delete();
     }
